added deleting history, ability to select distance and using unternational system units

+++ src/Route256.Week1.Homework.PriceCalculator.Api/Requests/V1/GoodProperties.cs
@@ -4,6 +4,6 @@ namespace Route256.Week1.Homework.PriceCalculator.Api.Requests.V1;
 /// Харектеристики товара
 /// </summary>
 public record GoodProperties(
-    int Height,
-    int Length,
-    int Width);
+    decimal Height,
+    decimal Length,
+    decimal Width);
