+++ src/Route256.Week1.Homework.PriceCalculator.Api/Requests/V2/GoodProperties.cs
@@ -4,7 +4,7 @@ namespace Route256.Week1.Homework.PriceCalculator.Api.Requests.V2;
 /// Харектеристики товара
 /// </summary>
 public record GoodProperties(
-    int Height,
-    int Length,
-    int Width,
-    int Weight);
+    decimal Height,
+    decimal Length,
+    decimal Width,
+    decimal Weight);
